--- opencti-platform/opencti-front/src/private/components/arsenal/malwares/Root.jsx
@@ -1,276 +1,0 @@
-import React, { Component } from 'react';
-import * as PropTypes from 'prop-types';
-import { Route, Routes, Link, Navigate } from 'react-router-dom';
-import { graphql } from 'react-relay';
-import Box from '@mui/material/Box';
-import Tabs from '@mui/material/Tabs';
-import Tab from '@mui/material/Tab';
-import * as R from 'ramda';
-import StixCoreObjectContentRoot from '../../common/stix_core_objects/StixCoreObjectContentRoot';
-import StixCoreObjectSimulationResult from '../../common/stix_core_objects/StixCoreObjectSimulationResult';
-import withRouter from '../../../../utils/compat-router/withRouter';
-import { QueryRenderer, requestSubscription } from '../../../../relay/environment';
-import Malware from './Malware';
-import MalwareKnowledge from './MalwareKnowledge';
-import StixDomainObjectHeader from '../../common/stix_domain_objects/StixDomainObjectHeader';
-import FileManager from '../../common/files/FileManager';
-import MalwarePopover from './MalwarePopover';
-import Loader from '../../../../components/Loader';
-import StixCoreObjectHistory from '../../common/stix_core_objects/StixCoreObjectHistory';
-import StixCoreObjectOrStixCoreRelationshipContainers from '../../common/containers/StixCoreObjectOrStixCoreRelationshipContainers';
-import StixCoreObjectKnowledgeBar from '../../common/stix_core_objects/StixCoreObjectKnowledgeBar';
-import ErrorNotFound from '../../../../components/ErrorNotFound';
-import inject18n from '../../../../components/i18n';
-import Breadcrumbs from '../../../../components/Breadcrumbs';
-import { getCurrentTab, getPaddingRight } from '../../../../utils/utils';
-
-const subscription = graphql`
-  subscription RootMalwareSubscription($id: ID!) {
-    stixDomainObject(id: $id) {
-      ... on Malware {
-        ...Malware_malware
-        ...MalwareEditionContainer_malware
-      }
-      ...FileImportViewer_entity
-      ...FileExportViewer_entity
-      ...FileExternalReferencesViewer_entity
-      ...WorkbenchFileViewer_entity
-      ...PictureManagementViewer_entity
-    }
-  }
-`;
-
-const malwareQuery = graphql`
-  query RootMalwareQuery($id: String!) {
-    malware(id: $id) {
-      id
-      standard_id
-      entity_type
-      name
-      aliases
-      x_opencti_graph_data
-      stixCoreObjectsDistribution(field: "entity_type", operation: count) {
-        label
-        value
-      }
-      ...Malware_malware
-      ...MalwareKnowledge_malware
-      ...FileImportViewer_entity
-      ...FileExportViewer_entity
-      ...FileExternalReferencesViewer_entity
-      ...WorkbenchFileViewer_entity
-      ...PictureManagementViewer_entity
-      ...StixCoreObjectContent_stixCoreObject
-    }
-    connectorsForImport {
-      ...FileManager_connectorsImport
-    }
-    connectorsForExport {
-      ...FileManager_connectorsExport
-    }
-  }
-`;
-
-class RootMalware extends Component {
-  constructor(props) {
-    super(props);
-    const {
-      params: { malwareId },
-    } = props;
-    this.sub = requestSubscription({
-      subscription,
-      variables: { id: malwareId },
-    });
-  }
-
-  componentWillUnmount() {
-    this.sub.dispose();
-  }
-
-  render() {
-    const {
-      t,
-      location,
-      params: { malwareId },
-    } = this.props;
-
-    const link = `/dashboard/arsenal/malwares/${malwareId}/knowledge`;
-    return (
-      <>
-        <QueryRenderer
-          query={malwareQuery}
-          variables={{ id: malwareId }}
-          render={({ props }) => {
-            if (props) {
-              if (props.malware) {
-                const { malware } = props;
-                const isOverview = location.pathname === `/dashboard/arsenal/malwares/${malwareId}`;
-                const paddingRight = getPaddingRight(location.pathname, malware.id, '/dashboard/arsenal/malwares');
-                return (
-                  <>
-                    <Routes>
-                      <Route
-                        path="/knowledge/*"
-                        element={
-                          <StixCoreObjectKnowledgeBar
-                            stixCoreObjectLink={link}
-                            availableSections={[
-                              'victimology',
-                              'threats',
-                              'threat_actors',
-                              'intrusion_sets',
-                              'campaigns',
-                              'incidents',
-                              'variants',
-                              'tools',
-                              'attack_patterns',
-                              'vulnerabilities',
-                              'indicators',
-                              'observables',
-                              'infrastructures',
-                              'sightings',
-                            ]}
-                            stixCoreObjectsDistribution={malware.stixCoreObjectsDistribution}
-                          />
-                        }
-                      />
-                    </Routes>
-                    <div style={{ paddingRight }}>
-                      <Breadcrumbs variant="object" elements={[
-                        { label: t('Arsenal') },
-                        { label: t('Malwares'), link: '/dashboard/arsenal/malwares' },
-                        { label: malware.name, current: true },
-                      ]}
-                      />
-                      <StixDomainObjectHeader
-                        entityType="Malware"
-                        stixDomainObject={malware}
-                        PopoverComponent={<MalwarePopover />}
-                        enableQuickSubscription={true}
-                      />
-                      <Box
-                        sx={{
-                          borderBottom: 1,
-                          borderColor: 'divider',
-                          marginBottom: 4,
-                        }}
-                      >
-                        <Tabs
-                          value={getCurrentTab(location.pathname, malware.id, '/dashboard/arsenal/malwares')}
-                        >
-                          <Tab
-                            component={Link}
-                            to={`/dashboard/arsenal/malwares/${malware.id}`}
-                            value={`/dashboard/arsenal/malwares/${malware.id}`}
-                            label={t('Overview')}
-                          />
-                          <Tab
-                            component={Link}
-                            to={`/dashboard/arsenal/malwares/${malware.id}/knowledge/overview`}
-                            value={`/dashboard/arsenal/malwares/${malware.id}/knowledge`}
-                            label={t('Knowledge')}
-                          />
-                          <Tab
-                            component={Link}
-                            to={`/dashboard/arsenal/malwares/${malware.id}/content`}
-                            value={`/dashboard/arsenal/malwares/${malware.id}/content`}
-                            label={t('Content')}
-                          />
-                          <Tab
-                            component={Link}
-                            to={`/dashboard/arsenal/malwares/${malware.id}/analyses`}
-                            value={`/dashboard/arsenal/malwares/${malware.id}/analyses`}
-                            label={t('Analyses')}
-                          />
-                          <Tab
-                            component={Link}
-                            to={`/dashboard/arsenal/malwares/${malware.id}/files`}
-                            value={`/dashboard/arsenal/malwares/${malware.id}/files`}
-                            label={t('Data')}
-                          />
-                          <Tab
-                            component={Link}
-                            to={`/dashboard/arsenal/malwares/${malware.id}/history`}
-                            value={`/dashboard/arsenal/malwares/${malware.id}/history`}
-                            label={t('History')}
-                          />
-                        </Tabs>
-                        {isOverview && (
-                          <StixCoreObjectSimulationResult id={malware.id} type="threat" />
-                        )}
-                      </Box>
-                      <Routes>
-                        <Route
-                          path="/"
-                          element={
-                            <Malware malware={malware} />}
-                        />
-                        <Route
-                          path="/knowledge"
-                          element={
-                            <Navigate
-                              replace={true}
-                              to={`/dashboard/arsenal/malwares/${malwareId}/knowledge/overview`}
-                            />}
-                        />
-                        <Route
-                          path="/knowledge/*"
-                          element={
-                            <MalwareKnowledge
-                              malware={malware}
-                            />}
-                        />
-                        <Route
-                          path="/content/*"
-                          element={
-                            <StixCoreObjectContentRoot
-                              stixCoreObject={malware}
-                            />
-                        }
-                        />
-                        <Route
-                          path="/analyses"
-                          element={
-                            <StixCoreObjectOrStixCoreRelationshipContainers
-                              stixDomainObjectOrStixCoreRelationship={malware}
-                            />}
-                        />
-                        <Route
-                          path="/files"
-                          element={
-                            <FileManager
-                              id={malwareId}
-                              connectorsImport={props.connectorsForImport}
-                              connectorsExport={props.connectorsForExport}
-                              entity={malware}
-                            />}
-                        />
-                        <Route
-                          path="/history"
-                          element={(
-                            <StixCoreObjectHistory
-                              stixCoreObjectId={malwareId}
-                            />
-                        )}
-                        />
-                      </Routes>
-                    </div>
-                  </>
-                );
-              }
-              return <ErrorNotFound />;
-            }
-            return <Loader />;
-          }}
-        />
-      </>
-    );
-  }
-}
-
-RootMalware.propTypes = {
-  children: PropTypes.node,
-  params: PropTypes.object,
-};
-
-export default R.compose(inject18n, withRouter)(RootMalware);
