+++ opencti-platform/opencti-front/src/private/components/arsenal/malwares/Root.tsx
@@ -3,39 +3,39 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-nocheck
 import React, { useMemo, Suspense } from 'react';
-import { Route, Routes, Link, Navigate, useLocation, useParams } from 'react-router-dom';
+import { Route, Routes, Link, Navigate, useParams, useLocation } from 'react-router-dom';
 import { graphql, useSubscription, usePreloadedQuery, PreloadedQuery } from 'react-relay';
-import { GraphQLSubscriptionConfig } from 'relay-runtime';
 import Box from '@mui/material/Box';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
 import useQueryLoading from 'src/utils/hooks/useQueryLoading';
+import { GraphQLSubscriptionConfig } from 'relay-runtime';
 import useForceUpdate from '@components/common/bulk/useForceUpdate';
+import BulkRelationDialogContainer from '@components/common/bulk/dialog/BulkRelationDialogContainer';
 import StixCoreObjectContentRoot from '../../common/stix_core_objects/StixCoreObjectContentRoot';
 import StixCoreObjectSimulationResult from '../../common/stix_core_objects/StixCoreObjectSimulationResult';
-import IntrusionSet from './IntrusionSet';
-import IntrusionSetKnowledge from './IntrusionSetKnowledge';
+import Malware from './Malware';
+import MalwareKnowledge from './MalwareKnowledge';
 import StixDomainObjectHeader from '../../common/stix_domain_objects/StixDomainObjectHeader';
 import FileManager from '../../common/files/FileManager';
-import IntrusionSetPopover from './IntrusionSetPopover';
+import MalwarePopover from './MalwarePopover';
 import Loader, { LoaderVariant } from '../../../../components/Loader';
 import StixCoreObjectHistory from '../../common/stix_core_objects/StixCoreObjectHistory';
 import StixCoreObjectOrStixCoreRelationshipContainers from '../../common/containers/StixCoreObjectOrStixCoreRelationshipContainers';
-import ErrorNotFound from '../../../../components/ErrorNotFound';
 import StixCoreObjectKnowledgeBar from '../../common/stix_core_objects/StixCoreObjectKnowledgeBar';
+import ErrorNotFound from '../../../../components/ErrorNotFound';
 import { useFormatter } from '../../../../components/i18n';
 import Breadcrumbs from '../../../../components/Breadcrumbs';
 import { getCurrentTab, getPaddingRight } from '../../../../utils/utils';
-import BulkRelationDialogContainer from '../../common/bulk/dialog/BulkRelationDialogContainer';
-import { RootIntrusionSetQuery } from './__generated__/RootIntrusionSetQuery.graphql';
-import { RootIntrusionSetSubscription } from './__generated__/RootIntrusionSetSubscription.graphql';
+import { RootMalwareQuery } from './__generated__/RootMalwareQuery.graphql';
+import { RootMalwareSubscription } from './__generated__/RootMalwareSubscription.graphql';
 
 const subscription = graphql`
-  subscription RootIntrusionSetSubscription($id: ID!) {
+  subscription RootMalwareSubscription($id: ID!) {
     stixDomainObject(id: $id) {
-      ... on IntrusionSet {
-        ...IntrusionSet_intrusionSet
-        ...IntrusionSetEditionContainer_intrusionSet
+      ... on Malware {
+        ...Malware_malware
+        ...MalwareEditionContainer_malware
       }
       ...FileImportViewer_entity
       ...FileExportViewer_entity
@@ -46,24 +46,21 @@ const subscription = graphql`
   }
 `;
 
-const intrusionSetQuery = graphql`
-  query RootIntrusionSetQuery($id: String!) {
-    intrusionSet(id: $id) {
+const malwareQuery = graphql`
+  query RootMalwareQuery($id: String!) {
+    malware(id: $id) {
       id
       standard_id
       entity_type
       name
       aliases
-      objectMarking {
-        id
-      }
       x_opencti_graph_data
       stixCoreObjectsDistribution(field: "entity_type", operation: count) {
         label
         value
       }
-      ...IntrusionSet_intrusionSet
-      ...IntrusionSetKnowledge_intrusionSet
+      ...Malware_malware
+      ...MalwareKnowledge_malware
       ...FileImportViewer_entity
       ...FileExportViewer_entity
       ...FileExternalReferencesViewer_entity
@@ -80,36 +77,36 @@ const intrusionSetQuery = graphql`
   }
 `;
 
-type RootIntrusionSetProps = {
-  intrusionSetId: string;
-  queryRef: PreloadedQuery<RootIntrusionSetQuery>;
+type RootMalwareProps = {
+  malwareId: string;
+  queryRef: PreloadedQuery<RootMalwareQuery>;
 };
-
-const RootIntrusionSet = ({ intrusionSetId, queryRef }: RootIntrusionSetProps) => {
-  const subConfig = useMemo<GraphQLSubscriptionConfig<RootIntrusionSetSubscription>>(() => ({
+const RootMalware = ({ queryRef, malwareId }: RootMalwareProps) => {
+  const subConfig = useMemo<GraphQLSubscriptionConfig<RootMalwareSubscription>>(() => ({
     subscription,
-    variables: { id: intrusionSetId },
-  }), [intrusionSetId]);
+    variables: { id: malwareId },
+  }), [malwareId]);
 
   const location = useLocation();
   const { t_i18n } = useFormatter();
   useSubscription(subConfig);
 
   const {
-    intrusionSet,
+    malware,
     connectorsForExport,
     connectorsForImport,
-  } = usePreloadedQuery<RootIntrusionSetQuery>(intrusionSetQuery, queryRef);
+  } = usePreloadedQuery<RootMalwareQuery>(malwareQuery, queryRef);
 
   const { forceUpdate, handleForceUpdate } = useForceUpdate();
 
-  const isOverview = location.pathname === `/dashboard/threats/intrusion_sets/${intrusionSetId}`;
-  const isKnowledge = location.pathname.startsWith(`/dashboard/threats/intrusion_sets/${intrusionSetId}/knowledge`);
-  const paddingRight = getPaddingRight(location.pathname, intrusionSetId, '/dashboard/threats/intrusion_sets');
-  const link = `/dashboard/threats/intrusion_sets/${intrusionSetId}/knowledge`;
+  const link = `/dashboard/arsenal/malwares/${malwareId}/knowledge`;
+  const isKnowledge = location.pathname.startsWith(`/dashboard/arsenal/malwares/${malwareId}/knowledge`);
+  const isOverview = location.pathname === `/dashboard/arsenal/malwares/${malwareId}`;
+  const paddingRight = getPaddingRight(location.pathname, malwareId, '/dashboard/arsenal/malwares');
+
   return (
     <>
-      {intrusionSet ? (
+      {malware ? (
         <>
           <Routes>
             <Route
@@ -119,39 +116,37 @@ const RootIntrusionSet = ({ intrusionSetId, queryRef }: RootIntrusionSetProps) =
                   stixCoreObjectLink={link}
                   availableSections={[
                     'victimology',
-                    'attribution',
+                    'threats',
+                    'threat_actors',
+                    'intrusion_sets',
                     'campaigns',
                     'incidents',
-                    'malwares',
-                    'attack_patterns',
+                    'variants',
                     'tools',
-                    'channels',
-                    'narratives',
+                    'attack_patterns',
                     'vulnerabilities',
                     'indicators',
                     'observables',
                     'infrastructures',
                     'sightings',
                   ]}
-                  stixCoreObjectsDistribution={intrusionSet.stixCoreObjectsDistribution}
-                  attribution={['Threat-Actor-Individual', 'Threat-Actor-Group']}
+                  stixCoreObjectsDistribution={malware.stixCoreObjectsDistribution}
                 />
               }
             />
           </Routes>
-          <div style={{ paddingRight }} data-testid="intrusionSet-details-page">
+          <div style={{ paddingRight }}>
             <Breadcrumbs variant="object" elements={[
-              { label: t_i18n('Threats') },
-              { label: t_i18n('Intrusion sets'), link: '/dashboard/threats/intrusion_sets' },
-              { label: intrusionSet.name, current: true },
+              { label: t_i18n('Arsenal') },
+              { label: t_i18n('Malwares'), link: '/dashboard/arsenal/malwares' },
+              { label: malware.name, current: true },
             ]}
             />
             <StixDomainObjectHeader
-              entityType="Intrusion-Set"
-              stixDomainObject={intrusionSet}
-              PopoverComponent={<IntrusionSetPopover />}
+              entityType="Malware"
+              stixDomainObject={malware}
+              PopoverComponent={<MalwarePopover />}
               enableQuickSubscription={true}
-              enableAskAi={true}
             />
             <Box
               sx={{
@@ -161,75 +156,80 @@ const RootIntrusionSet = ({ intrusionSetId, queryRef }: RootIntrusionSetProps) =
               }}
             >
               <Tabs
-                value={getCurrentTab(location.pathname, intrusionSet.id, '/dashboard/threats/intrusion_sets')}
+                value={getCurrentTab(location.pathname, malware.id, '/dashboard/arsenal/malwares')}
               >
                 <Tab
                   component={Link}
-                  to={`/dashboard/threats/intrusion_sets/${intrusionSet.id}`}
-                  value={`/dashboard/threats/intrusion_sets/${intrusionSet.id}`}
+                  to={`/dashboard/arsenal/malwares/${malware.id}`}
+                  value={`/dashboard/arsenal/malwares/${malware.id}`}
                   label={t_i18n('Overview')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/threats/intrusion_sets/${intrusionSet.id}/knowledge/overview`}
-                  value={`/dashboard/threats/intrusion_sets/${intrusionSet.id}/knowledge`}
+                  to={`/dashboard/arsenal/malwares/${malware.id}/knowledge/overview`}
+                  value={`/dashboard/arsenal/malwares/${malware.id}/knowledge`}
                   label={t_i18n('Knowledge')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/threats/intrusion_sets/${intrusionSet.id}/content`}
-                  value={`/dashboard/threats/intrusion_sets/${intrusionSet.id}/content`}
+                  to={`/dashboard/arsenal/malwares/${malware.id}/content`}
+                  value={`/dashboard/arsenal/malwares/${malware.id}/content`}
                   label={t_i18n('Content')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/threats/intrusion_sets/${intrusionSet.id}/analyses`}
-                  value={`/dashboard/threats/intrusion_sets/${intrusionSet.id}/analyses`}
+                  to={`/dashboard/arsenal/malwares/${malware.id}/analyses`}
+                  value={`/dashboard/arsenal/malwares/${malware.id}/analyses`}
                   label={t_i18n('Analyses')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/threats/intrusion_sets/${intrusionSet.id}/files`}
-                  value={`/dashboard/threats/intrusion_sets/${intrusionSet.id}/files`}
+                  to={`/dashboard/arsenal/malwares/${malware.id}/files`}
+                  value={`/dashboard/arsenal/malwares/${malware.id}/files`}
                   label={t_i18n('Data')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/threats/intrusion_sets/${intrusionSet.id}/history`}
-                  value={`/dashboard/threats/intrusion_sets/${intrusionSet.id}/history`}
+                  to={`/dashboard/arsenal/malwares/${malware.id}/history`}
+                  value={`/dashboard/arsenal/malwares/${malware.id}/history`}
                   label={t_i18n('History')}
                 />
               </Tabs>
               {isKnowledge && (
                 <BulkRelationDialogContainer
-                  stixDomainObjectId={intrusionSet.id}
-                  stixDomainObjectName={intrusionSet.name}
-                  stixDomainObjectType="Intrusion-Set"
+                  stixDomainObjectId={malware.id}
+                  stixDomainObjectName={malware.name}
+                  stixDomainObjectType="Malware"
                   handleRefetch={handleForceUpdate}
                 />
               )}
               {isOverview && (
-                <StixCoreObjectSimulationResult id={intrusionSet.id} type="threat" />
+                <StixCoreObjectSimulationResult id={malware.id} type="threat" />
               )}
             </Box>
             <Routes>
               <Route
                 path="/"
                 element={
-                  <IntrusionSet intrusionSet={intrusionSet} />
+                  <Malware malware={malware} />
                 }
               />
               <Route
                 path="/knowledge"
                 element={
-                  <Navigate to={`/dashboard/threats/intrusion_sets/${intrusionSetId}/knowledge/overview`} replace={true} />
+                  <Navigate
+                    replace={true}
+                    to={`/dashboard/arsenal/malwares/${malwareId}/knowledge/overview`}
+                  />
                 }
               />
               <Route
                 path="/knowledge/*"
                 element={
-                  <div data-testid="instrusionSet-knowledge" key={forceUpdate}>
-                    <IntrusionSetKnowledge intrusionSet={intrusionSet} />
+                  <div key={forceUpdate}>
+                    <MalwareKnowledge
+                      malware={malware}
+                    />
                   </div>
                 }
               />
@@ -237,31 +237,35 @@ const RootIntrusionSet = ({ intrusionSetId, queryRef }: RootIntrusionSetProps) =
                 path="/content/*"
                 element={
                   <StixCoreObjectContentRoot
-                    stixCoreObject={intrusionSet}
+                    stixCoreObject={malware}
                   />
                 }
               />
               <Route
                 path="/analyses"
                 element={
-                  <StixCoreObjectOrStixCoreRelationshipContainers stixDomainObjectOrStixCoreRelationship={intrusionSet} />
+                  <StixCoreObjectOrStixCoreRelationshipContainers
+                    stixDomainObjectOrStixCoreRelationship={malware}
+                  />
                 }
               />
               <Route
                 path="/files"
                 element={
                   <FileManager
-                    id={intrusionSetId}
+                    id={malwareId}
                     connectorsImport={connectorsForImport}
                     connectorsExport={connectorsForExport}
-                    entity={intrusionSet}
+                    entity={malware}
                   />
                 }
               />
               <Route
                 path="/history"
                 element={
-                  <StixCoreObjectHistory stixCoreObjectId={intrusionSetId} />
+                  <StixCoreObjectHistory
+                    stixCoreObjectId={malwareId}
+                  />
                 }
               />
             </Routes>
@@ -275,16 +279,16 @@ const RootIntrusionSet = ({ intrusionSetId, queryRef }: RootIntrusionSetProps) =
 };
 
 const Root = () => {
-  const { intrusionSetId } = useParams();
-  const queryRef = useQueryLoading<RootIntrusionSetQuery>(intrusionSetQuery, {
-    id: intrusionSetId ?? '',
+  const { malwareId } = useParams();
+  const queryRef = useQueryLoading<RootMalwareQuery>(malwareQuery, {
+    id: malwareId ?? '',
   });
 
   return (
     <>
       {queryRef && (
         <Suspense fallback={<Loader variant={LoaderVariant.container} />}>
-          <RootIntrusionSet queryRef={queryRef} intrusionSetId={intrusionSetId} />
+          <RootMalware queryRef={queryRef} malwareId={malwareId} />
         </Suspense>
       )}
     </>
